refactoring, rid out view from viewmodel
