+++ app/src/main/res/layout/activity_github.xml
@@ -9,6 +9,10 @@
         <variable
             name="optionID"
             type="Integer" />
+
+        <variable
+            name="viewModel"
+            type="com.example.answer.github.viewmodel.GithubViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
